refactor trace console e2e support
