+++ crates/rulemorph_ui/ui/e2e/trace_console.spec.ts
@@ -1,497 +1,17 @@
 import { test, expect } from "@playwright/test";
-import { spawn } from "child_process";
-import { mkdtempSync, writeFileSync, mkdirSync, existsSync, readFileSync } from "fs";
+import { mkdtempSync, writeFileSync, mkdirSync, readFileSync } from "fs";
 import { tmpdir } from "os";
 import path from "path";
-import net from "net";
-import { fileURLToPath } from "url";
-
-const repoRoot = path.resolve(
-  path.dirname(fileURLToPath(import.meta.url)),
-  "../../../.."
-);
-const uiDist = path.resolve(repoRoot, "crates/rulemorph_ui/ui/dist");
-const apiRulesDir = path.resolve(repoRoot, "assets/api_rules");
-
-function crc32(buffer: Buffer): number {
-  let crc = 0xffffffff;
-  for (const byte of buffer) {
-    crc ^= byte;
-    for (let bit = 0; bit < 8; bit += 1) {
-      crc = (crc >>> 1) ^ (crc & 1 ? 0xedb88320 : 0);
-    }
-  }
-  return (crc ^ 0xffffffff) >>> 0;
-}
-
-function dosDateTime(date = new Date("2026-02-03T00:00:00Z")) {
-  const year = Math.max(1980, date.getUTCFullYear());
-  const dosTime =
-    (date.getUTCHours() << 11) | (date.getUTCMinutes() << 5) | Math.floor(date.getUTCSeconds() / 2);
-  const dosDate = ((year - 1980) << 9) | ((date.getUTCMonth() + 1) << 5) | date.getUTCDate();
-  return { dosDate, dosTime };
-}
-
-function createStoredZip(zipPath: string, entries: Array<{ name: string; data: Buffer }>) {
-  const localParts: Buffer[] = [];
-  const centralParts: Buffer[] = [];
-  let offset = 0;
-  const { dosDate, dosTime } = dosDateTime();
-
-  for (const entry of entries) {
-    const name = Buffer.from(entry.name.replace(/\\/g, "/"));
-    const data = entry.data;
-    const checksum = crc32(data);
-
-    const local = Buffer.alloc(30);
-    local.writeUInt32LE(0x04034b50, 0);
-    local.writeUInt16LE(20, 4);
-    local.writeUInt16LE(0, 6);
-    local.writeUInt16LE(0, 8);
-    local.writeUInt16LE(dosTime, 10);
-    local.writeUInt16LE(dosDate, 12);
-    local.writeUInt32LE(checksum, 14);
-    local.writeUInt32LE(data.length, 18);
-    local.writeUInt32LE(data.length, 22);
-    local.writeUInt16LE(name.length, 26);
-    local.writeUInt16LE(0, 28);
-    localParts.push(local, name, data);
-
-    const central = Buffer.alloc(46);
-    central.writeUInt32LE(0x02014b50, 0);
-    central.writeUInt16LE(20, 4);
-    central.writeUInt16LE(20, 6);
-    central.writeUInt16LE(0, 8);
-    central.writeUInt16LE(0, 10);
-    central.writeUInt16LE(dosTime, 12);
-    central.writeUInt16LE(dosDate, 14);
-    central.writeUInt32LE(checksum, 16);
-    central.writeUInt32LE(data.length, 20);
-    central.writeUInt32LE(data.length, 24);
-    central.writeUInt16LE(name.length, 28);
-    central.writeUInt16LE(0, 30);
-    central.writeUInt16LE(0, 32);
-    central.writeUInt16LE(0, 34);
-    central.writeUInt16LE(0, 36);
-    central.writeUInt32LE(0, 38);
-    central.writeUInt32LE(offset, 42);
-    centralParts.push(central, name);
-
-    offset += local.length + name.length + data.length;
-  }
-
-  const centralSize = centralParts.reduce((sum, part) => sum + part.length, 0);
-  const end = Buffer.alloc(22);
-  end.writeUInt32LE(0x06054b50, 0);
-  end.writeUInt16LE(0, 4);
-  end.writeUInt16LE(0, 6);
-  end.writeUInt16LE(entries.length, 8);
-  end.writeUInt16LE(entries.length, 10);
-  end.writeUInt32LE(centralSize, 12);
-  end.writeUInt32LE(offset, 16);
-  end.writeUInt16LE(0, 20);
-
-  writeFileSync(zipPath, Buffer.concat([...localParts, ...centralParts, end]));
-}
-
-function getAvailablePort(): Promise<number> {
-  return new Promise((resolve, reject) => {
-    const server = net.createServer();
-    server.unref();
-    server.on("error", reject);
-    server.listen(0, () => {
-      const address = server.address();
-      if (!address || typeof address === "string") {
-        server.close(() => reject(new Error("failed to acquire port")));
-        return;
-      }
-      const { port } = address;
-      server.close(() => resolve(port));
-    });
-  });
-}
-
-function writeDemoTrace(dataDir: string) {
-  const traceDir = path.join(dataDir, "traces", "2026", "02", "03", "demo-001");
-  mkdirSync(traceDir, { recursive: true });
-  writeFileSync(
-    path.join(traceDir, "records-0001.ndjson"),
-    "{\"index\":0,\"status\":\"ok\",\"duration_us\":1000,\"input\":{\"foo\":1},\"output\":{\"bar\":2}}\n"
-  );
-  writeFileSync(
-    path.join(traceDir, "nodes-0001.ndjson"),
-    "{\"record_index\":0,\"id\":\"n1\",\"kind\":\"mappings\",\"label\":\"mappings[0]\",\"status\":\"ok\",\"input\":{\"foo\":1},\"output\":{\"bar\":2},\"children\":[{\"id\":\"n1.1\",\"kind\":\"op\",\"label\":\"add\",\"status\":\"ok\",\"input\":1,\"output\":2,\"pipe_value\":1,\"args\":[1],\"meta\":{\"op\":\"add\"}}]}\n"
-  );
-  writeFileSync(
-    path.join(traceDir, "finalize.json"),
-    "{\"status\":\"ok\",\"input\":[{\"bar\":2}],\"output\":[{\"bar\":2}],\"duration_us\":100,\"nodes\":[{\"id\":\"op-wrap\",\"kind\":\"op\",\"label\":\"wrap\",\"status\":\"ok\",\"meta\":{\"op\":\"wrap\"},\"args\":{\"wrap\":{\"key\":\"value\"}}}]}"
-  );
-  writeFileSync(
-    path.join(traceDir, "trace.json"),
-    JSON.stringify(
-      {
-        trace_schema_version: 1,
-        trace_id: "demo-001",
-        timestamp: "2026-02-03T00:00:00Z",
-        status: "ok",
-        rule: {
-          type: "normal",
-          name: "demo",
-          path: "rules/demo.yaml",
-          version: 2
-        },
-        input_format: "json",
-        summary: {
-          record_total: 1,
-          record_success: 1,
-          record_failed: 0,
-          duration_us: 1000
-        },
-        max_chunk_bytes_uncompressed: 1048576,
-        detail: {
-          layout: "records_nodes_split",
-          status: "full",
-          reason: [],
-          records: [
-            {
-              path: "records-0001.ndjson",
-              format: "ndjson",
-              compression: "none",
-              record_start: 0,
-              record_end: 0
-            }
-          ],
-          nodes: [
-            {
-              path: "nodes-0001.ndjson",
-              format: "ndjson",
-              compression: "none",
-              node_start: 0,
-              node_end: 0
-            }
-          ],
-          finalize: {
-            path: "finalize.json",
-            format: "json",
-            compression: "none"
-          }
-        }
-      },
-      null,
-      2
-    )
-  );
-}
-
-function writeInlineNodesTrace(dataDir: string) {
-  const traceDir = path.join(dataDir, "traces", "2026", "02", "03", "inline-001");
-  mkdirSync(traceDir, { recursive: true });
-  writeFileSync(
-    path.join(traceDir, "records-0001.ndjson"),
-    "{\"index\":0,\"status\":\"ok\",\"duration_us\":500,\"nodes\":{\"id\":\"inline-0\",\"kind\":\"mappings\",\"label\":\"inline node\",\"status\":\"ok\",\"input\":{\"foo\":1},\"output\":{\"bar\":2}}}\n"
-  );
-  writeFileSync(
-    path.join(traceDir, "trace.json"),
-    JSON.stringify(
-      {
-        trace_schema_version: 1,
-        trace_id: "inline-001",
-        timestamp: "2026-02-03T00:05:00Z",
-        status: "ok",
-        rule: {
-          type: "normal",
-          name: "inline",
-          path: "rules/inline.yaml",
-          version: 2
-        },
-        input_format: "json",
-        summary: {
-          record_total: 1,
-          record_success: 1,
-          record_failed: 0,
-          duration_us: 500
-        },
-        max_chunk_bytes_uncompressed: 1048576,
-        detail: {
-          layout: "records_inline",
-          status: "full",
-          reason: [],
-          records: [
-            {
-              path: "records-0001.ndjson",
-              format: "ndjson",
-              compression: "none",
-              record_start: 0,
-              record_end: 0
-            }
-          ],
-          nodes: []
-        }
-      },
-      null,
-      2
-    )
-  );
-}
-
-function writeLegacyInlineNodesTrace(dataDir: string) {
-  const traceDir = path.join(dataDir, "traces", "2026", "02", "03", "legacy-inline-001");
-  mkdirSync(traceDir, { recursive: true });
-  writeFileSync(
-    path.join(traceDir, "trace.json"),
-    JSON.stringify(
-      {
-        trace_id: "legacy-inline-001",
-        timestamp: "2026-02-03T00:08:00Z",
-        status: "ok",
-        rule: {
-          type: "normal",
-          name: "legacy inline",
-          path: "rules/legacy-inline.yaml",
-          version: 2
-        },
-        records: [
-          {
-            index: 0,
-            status: "ok",
-            duration_us: 450,
-            nodes: {
-              id: "legacy-inline-0",
-              kind: "mappings",
-              label: "legacy inline",
-              status: "ok",
-              input: { foo: 1 },
-              output: { bar: 2 }
-            }
-          }
-        ]
-      },
-      null,
-      2
-    )
-  );
-}
-
-function writeBasicTrace(dataDir: string) {
-  const traceDir = path.join(dataDir, "traces", "2026", "02", "03", "basic-001");
-  mkdirSync(traceDir, { recursive: true });
-  writeFileSync(
-    path.join(traceDir, "trace.json"),
-    JSON.stringify(
-      {
-        trace_schema_version: 1,
-        trace_id: "basic-001",
-        timestamp: "2026-02-03T00:10:00Z",
-        status: "ok",
-        rule: {
-          type: "normal",
-          name: "basic",
-          path: "rules/basic.yaml",
-          version: 2
-        },
-        input_format: "json",
-        summary: {
-          record_total: 1,
-          record_success: 1,
-          record_failed: 0,
-          duration_us: 900
-        },
-        max_chunk_bytes_uncompressed: 1048576,
-        detail: {
-          layout: "records_nodes_split",
-          status: "basic",
-          reason: ["budget_exceeded"],
-          records: [],
-          nodes: []
-        }
-      },
-      null,
-      2
-    )
-  );
-}
-
-function writeErrorTrace(dataDir: string) {
-  const traceDir = path.join(dataDir, "traces", "2026", "02", "03", "error-001");
-  mkdirSync(traceDir, { recursive: true });
-  writeFileSync(
-    path.join(traceDir, "trace.json"),
-    JSON.stringify(
-      {
-        trace_schema_version: 1,
-        trace_id: "error-001",
-        timestamp: "2026-02-03T00:15:00Z",
-        status: "error",
-        rule: {
-          type: "normal",
-          name: "error",
-          path: "rules/error.yaml",
-          version: 2
-        },
-        input_format: "json",
-        summary: {
-          record_total: 1,
-          record_success: 0,
-          record_failed: 1,
-          duration_us: 1200
-        },
-        max_chunk_bytes_uncompressed: 1048576,
-        detail: {
-          layout: "records_nodes_split",
-          status: "basic",
-          reason: [],
-          records: [],
-          nodes: []
-        }
-      },
-      null,
-      2
-    )
-  );
-}
-
-function writeBrokenTrace(dataDir: string) {
-  const traceDir = path.join(dataDir, "traces", "2026", "02", "03", "broken-001");
-  mkdirSync(traceDir, { recursive: true });
-  writeFileSync(
-    path.join(traceDir, "trace.json"),
-    JSON.stringify(
-      {
-        trace_schema_version: 1,
-        trace_id: "broken-001",
-        timestamp: "2026-02-03T00:20:00Z",
-        status: "ok",
-        rule: {
-          type: "normal",
-          name: "broken",
-          path: "rules/broken.yaml",
-          version: 2
-        },
-        input_format: "json",
-        summary: {
-          record_total: 1,
-          record_success: 1,
-          record_failed: 0,
-          duration_us: 700
-        },
-        max_chunk_bytes_uncompressed: 1048576,
-        detail: {
-          layout: "records_nodes_split",
-          status: "full",
-          reason: [],
-          records: [
-            {
-              path: "records-0001.ndjson",
-              format: "ndjson",
-              compression: "none",
-              record_start: 0,
-              record_end: 0
-            }
-          ],
-          nodes: []
-        }
-      },
-      null,
-      2
-    )
-  );
-}
-
-async function startServer(dataDir: string, port: number) {
-  return startServerWithOptions(dataDir, port, {
-    allowUnauthInternal: true
-  });
-}
-
-type ServerOptions = {
-  apiKey?: string;
-  tenantId?: string;
-  internalApiKey?: string;
-  allowUnauthInternal?: boolean;
-};
-
-async function startServerWithOptions(dataDir: string, port: number, options: ServerOptions) {
-  if (!existsSync(uiDist)) {
-    throw new Error(`UI dist not found at ${uiDist}. Run npm --prefix crates/rulemorph_ui/ui run build first.`);
-  }
-  if (!existsSync(apiRulesDir)) {
-    throw new Error(`api rules not found at ${apiRulesDir}.`);
-  }
-  const args = [
-    "run",
-    "-p",
-    "rulemorph_server",
-    "--",
-    "--api-mode",
-    "rules",
-    "--rules-dir",
-    apiRulesDir,
-    "--data-dir",
-    dataDir,
-    "--ui-dir",
-    uiDist,
-    "--ssrf-allow-private",
-    "--port",
-    String(port)
-  ];
-  if (options.allowUnauthInternal ?? false) {
-    args.push("--allow-unauth-internal");
-  }
-  if (options.apiKey) {
-    args.push("--api-key", options.apiKey);
-  }
-  if (options.tenantId) {
-    args.push("--tenant-id", options.tenantId);
-  }
-  if (options.internalApiKey) {
-    args.push("--internal-api-key", options.internalApiKey);
-  }
-  if (options.apiKey) {
-    args.push("--ssrf-allow-any");
-  }
-
-  const child = spawn(
-    "cargo",
-    args,
-    {
-      cwd: repoRoot,
-      env: { ...process.env, RUST_LOG: "info" },
-      stdio: ["ignore", "pipe", "pipe"]
-    }
-  );
-
-  await new Promise<void>((resolve, reject) => {
-    let settled = false;
-    const timeout = setTimeout(() => {
-      settled = true;
-      reject(new Error("server did not start in time"));
-    }, 20_000);
-    const cleanup = () => {
-      clearTimeout(timeout);
-      child.stdout?.off("data", onData);
-      child.stderr?.off("data", onData);
-    };
-    const onData = (data: Buffer) => {
-      const text = data.toString();
-      if (text.includes("listening on 127.0.0.1")) {
-        settled = true;
-        cleanup();
-        resolve();
-      }
-    };
-    child.stdout?.on("data", onData);
-    child.stderr?.on("data", onData);
-    child.on("exit", (code) => {
-      if (!settled) {
-        settled = true;
-        cleanup();
-        reject(new Error(`server exited early: ${code}`));
-      }
-    });
-  });
-
-  return child;
-}
+import { getAvailablePort, startServer, startServerWithOptions } from "./support/server";
+import {
+  writeBasicTrace,
+  writeBrokenTrace,
+  writeDemoTrace,
+  writeErrorTrace,
+  writeInlineNodesTrace,
+  writeLegacyInlineNodesTrace
+} from "./support/traceFixtures";
+import { createStoredZip } from "./support/zip";
 
 test.describe.configure({ mode: "serial" });
